+++ access/website/scripts/postgres/create-db.sh
@@ -1,1 +1,1 @@
-docker run --net access_default postgres:9.6.2 createdb -h shinedb -U postgres shine
+docker run --net website_default postgres:9.6.2 createdb -h shinedb -U postgres shine