+++ access/website/scripts/postgres/create-user.sh
@@ -1,2 +1,2 @@
-docker run --net access_default -v $PWD/setup_user.sql:/tmp/setup_user.sql postgres:9.6.2 psql -h shinedb -v -U postgres -d shine -f /tmp/setup_user.sql
+docker run --net website_default -v $PWD/setup_user.sql:/tmp/setup_user.sql postgres:9.6.2 psql -h shinedb -v -U postgres -d shine -f /tmp/setup_user.sql